fix: MET Norway float humidity values crashing JSON decode
MET Norway returns relative_humidity as a float (e.g. 73.4), but the
Swift Hourly model typed it as [Int?]? — causing typeMismatch on every
decode and the "Veri, olmadığı için okunamadı" error shown to the user.

- OpenMeteoModels: [Int?]? → [Double?]? for Hourly.relativeHumidity2m
- WeatherFusion: replace intRoundedCol (now unused) with doubleCol
- WeatherMapping: drop redundant Double(Int) cast
- Worker: guard against null current when MET Norway timeseries is empty

+++ Praeventus.swiftpm/OpenMeteoModels.swift
@@ -59,7 +59,7 @@ struct ForecastResponse: Codable, Equatable {
         let windSpeed10m: [Double?]?
         let windDirection10m: [Int?]?
         let windGusts10m: [Double?]?
-        let relativeHumidity2m: [Int?]?
+        let relativeHumidity2m: [Double?]?
         let dewPoint2m: [Double?]?
         let visibility: [Double?]?
